Converted int values to longs.

+++ src/main/scala/atlas/Cell.scala
@@ -8,10 +8,10 @@ case class Cell(
   position: Vector2,
 
   // The food available in this cell.
-  food: Int = 100,
+  food: Long = 100,
 
   // The water available in this cell.
-  water: Int = 100,
+  water: Long = 100,
 
   // The set of organisms occupying this cell.
   organisms: Set[Organism] = Set.empty
